pyproject.toml
[tool.poetry]
name = "bot-pet"
version = "0.1.0"
description = "aiogram bot"
authors = ["bbt-t github.com/bbt-t"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.16"
APScheduler = "^3.8.1"
python-dotenv = "^0.19.2"
loguru = "^0.5.3"
vosk = "^0.3.32"
SQLAlchemy = "^1.4.27"
aioredis = "^2.0.0"
asyncpg = "^0.25.0"
PGPy = "^0.5.4"
aiosqlite = "^0.17.0"
beautifulsoup4 = "^4.10.0"
lxml = "^4.7.1"
uvloop = "^0.16.0"
ujson = "^5.1.0"
cchardet = "^2.1.7"
aiodns = "^3.0.0"
aiofiles = "^0.8.0"
pytest-asyncio = "^0.17.2"
pydantic = "^1.9.0"
gTTS = "^2.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
